requirements.txt
mteb==1.19.2
einops==0.8.0
numpy==2.1.2
tiktoken==0.8.0
torch==2.5.1
ml_dtypes==0.5.0
qdrant-client==1.12.1
scikit-learn==1.5.2
PyYAML==6.0.2
pandas==2.2.3
seaborn==0.13.2
umap-learn==0.5.7
